create package-button and move files
